refactor: update auth flow and pages
