+++ frontend/src/components/LandingPage.tsx
@@ -1,0 +1,114 @@
+export default function LandingPage() {
+  return (
+    <div className="min-h-screen bg-cream text-dark-green">
+      {/* Hero Section */}
+      <section className="px-6 py-20">
+        <div className="max-w-7xl mx-auto text-center">
+          <h1 className="text-4xl md:text-6xl font-bold text-dark-green mb-6">
+            Collaborative Document Editing
+            <span className="block text-3xl md:text-4xl font-normal text-gray-600 mt-2">
+              Built for Teams
+            </span>
+          </h1>
+          <p className="text-xl text-gray-600 mb-8 max-w-3xl mx-auto">
+            Create, edit, and collaborate on documents in real-time with your
+            team. Experience seamless collaboration with our powerful editor
+            powered by TipTap.
+          </p>
+        </div>
+      </section>
+
+      {/* Features Section */}
+      <section className="px-6 py-20 bg-white">
+        <div className="max-w-7xl mx-auto">
+          <h2 className="text-3xl font-bold text-dark-green text-center mb-12">
+            Why Choose Agent Docs?
+          </h2>
+          <div className="grid md:grid-cols-3 gap-8">
+            <div className="text-center">
+              <div className="w-16 h-16 bg-light-green rounded-full flex items-center justify-center mx-auto mb-4">
+                <svg
+                  className="w-8 h-8 text-dark-green"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M12 6v6m0 0v6m0-6h6m-6 0H6"
+                  />
+                </svg>
+              </div>
+              <h3 className="text-xl font-semibold text-dark-green mb-2">
+                Real-time Collaboration
+              </h3>
+              <p className="text-gray-600">
+                Work together with your team in real-time. See changes as they
+                happen with live cursors and instant updates.
+              </p>
+            </div>
+            <div className="text-center">
+              <div className="w-16 h-16 bg-light-green rounded-full flex items-center justify-center mx-auto mb-4">
+                <svg
+                  className="w-8 h-8 text-dark-green"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
+                  />
+                </svg>
+              </div>
+              <h3 className="text-xl font-semibold text-dark-green mb-2">
+                Rich Text Editing
+              </h3>
+              <p className="text-gray-600">
+                Powerful editor with formatting options, lists, headings, and
+                more. Create professional documents with ease.
+              </p>
+            </div>
+            <div className="text-center">
+              <div className="w-16 h-16 bg-light-green rounded-full flex items-center justify-center mx-auto mb-4">
+                <svg
+                  className="w-8 h-8 text-dark-green"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"
+                  />
+                </svg>
+              </div>
+              <h3 className="text-xl font-semibold text-dark-green mb-2">
+                Secure & Private
+              </h3>
+              <p className="text-gray-600">
+                Your documents are secure and private. Built with modern
+                authentication and data protection.
+              </p>
+            </div>
+          </div>
+        </div>
+      </section>
+
+      {/* Footer */}
+      <footer className="border-t border-gray-200 px-6 py-8">
+        <div className="max-w-7xl mx-auto text-center">
+          <p className="text-gray-600">
+            © 2024 Agent Docs. Built with Next.js, FastAPI, and TipTap.
+          </p>
+        </div>
+      </footer>
+    </div>
+  );
+}
