+++ frontend/src/hooks/useTiptapEditor.ts
@@ -22,17 +22,20 @@ export const useTiptapEditor = ({
   const providerRef = useRef<HocuspocusProvider | null>(null);
   const ydocRef = useRef<Y.Doc | null>(null);
 
+  // Temporarily disable collaboration to fix WebSocket issues
+  const enableCollaboration = false; // documentId && token && typeof window !== "undefined";
+
   const editor = useEditor({
     extensions: [
       ...extensions,
-      documentId && token && typeof window !== "undefined"
+      ...(enableCollaboration
         ? [
             Collaboration.configure({
               document: (() => {
                 const ydoc = new Y.Doc();
                 ydocRef.current = ydoc;
                 providerRef.current = new HocuspocusProvider({
-                  name: documentId,
+                  name: documentId!,
                   url: process.env.NEXT_PUBLIC_TIPTAP_COLLAB_URL!,
                   token,
                   document: ydoc,
@@ -45,9 +48,9 @@ export const useTiptapEditor = ({
               user: user || { name: "Anonymous", color: "#18A0FB" },
             }),
           ]
-        : [],
-    ].flat(),
-    content: documentId ? undefined : DEFAULT_EDITOR_CONTENT, // Only use default content if no documentId (collaboration)
+        : []),
+    ],
+    content: enableCollaboration ? undefined : DEFAULT_EDITOR_CONTENT,
     editable: true,
     autofocus: true,
     immediatelyRender: false,
@@ -55,8 +58,14 @@ export const useTiptapEditor = ({
 
   useEffect(() => {
     return () => {
-      providerRef.current?.destroy();
-      ydocRef.current?.destroy();
+      if (providerRef.current) {
+        providerRef.current.destroy();
+        providerRef.current = null;
+      }
+      if (ydocRef.current) {
+        ydocRef.current.destroy();
+        ydocRef.current = null;
+      }
     };
   }, []);
